refactor: 検索ヘルパー関数を shared.rs へ共通化
escape_like_pattern（dividend/domestic_stock/mutualfund/asset_balance）と
parse_date_param・year_to_range・parse_year_month_range（dividend/domestic_stock/
mutualfund）が完全に同一内容で複数ファイルに重複していたため、
backend/src/services/shared.rs へ集約する。

LIKE エスケープはセキュリティ関連ロジックであり、修正漏れのリスクが
特に高かった。挙動は変更していない（純粋な移動・重複排除）。
