+++ backend/src/services/shared.rs
@@ -1,5 +1,6 @@
 use crate::errors::ApiError;
 use crate::models::common::BulkCreateResponse;
+use chrono::NaiveDate;
 use sqlx::postgres::PgQueryResult;
 use sqlx::PgPool;
 use std::time::Instant;
@@ -105,10 +106,104 @@ pub async fn delete_all_for_user(
     info!("[{}.delete_all] 完了: {}件削除", domain, deleted);
     Ok(deleted)
 }
+
+pub fn parse_date_param(field: &str, value: &str) -> Result<NaiveDate, ApiError> {
+    NaiveDate::parse_from_str(value, "%Y-%m-%d")
+        .map_err(|_| ApiError::ValidationError(format!("{field} の形式が不正です（YYYY-MM-DD）")))
+}
+
+pub fn year_to_range(year: i32) -> Result<(NaiveDate, NaiveDate), ApiError> {
+    let invalid = || ApiError::ValidationError("year の値が不正です".to_string());
+    let start = NaiveDate::from_ymd_opt(year, 1, 1).ok_or_else(invalid)?;
+    let end = NaiveDate::from_ymd_opt(year + 1, 1, 1).ok_or_else(invalid)?;
+    Ok((start, end))
+}
+
+pub fn parse_year_month_range(value: &str) -> Result<(NaiveDate, NaiveDate), ApiError> {
+    let invalid =
+        || ApiError::ValidationError("year_month の形式が不正です（YYYY-MM）".to_string());
+    let (year_str, month_str) = value.split_once('-').ok_or_else(invalid)?;
+    let year: i32 = year_str.parse().map_err(|_| invalid())?;
+    let month: u32 = month_str.parse().map_err(|_| invalid())?;
+    let start = NaiveDate::from_ymd_opt(year, month, 1).ok_or_else(invalid)?;
+    let end = if month == 12 {
+        NaiveDate::from_ymd_opt(year + 1, 1, 1)
+    } else {
+        NaiveDate::from_ymd_opt(year, month + 1, 1)
+    }
+    .ok_or_else(invalid)?;
+    Ok((start, end))
+}
+
+/// ILIKE の wildcard 文字（%, _, \）をリテラル扱いにエスケープしてから前後を % で囲む
+pub fn escape_like_pattern(token: &str) -> String {
+    let escaped = token
+        .replace('\\', "\\\\")
+        .replace('%', "\\%")
+        .replace('_', "\\_");
+    format!("%{escaped}%")
+}
+
 #[cfg(test)]
 mod tests {
-    use super::{user_ids_for_bulk_insert, BulkTimer};
+    use super::{
+        escape_like_pattern, parse_date_param, parse_year_month_range, user_ids_for_bulk_insert,
+        year_to_range, BulkTimer,
+    };
+    use crate::errors::ApiError;
+    use chrono::NaiveDate;
     use uuid::Uuid;
+
+    #[test]
+    fn test_year_to_range_produces_year_boundaries() {
+        let (start, end) = year_to_range(2026).expect("2026年は有効な範囲");
+        assert_eq!(start, NaiveDate::from_ymd_opt(2026, 1, 1).unwrap());
+        assert_eq!(end, NaiveDate::from_ymd_opt(2027, 1, 1).unwrap());
+    }
+
+    #[test]
+    fn test_parse_year_month_range_handles_december_wrap_and_invalid_values() {
+        let (start, end) = parse_year_month_range("2026-06").expect("2026-06 は有効");
+        assert_eq!(start, NaiveDate::from_ymd_opt(2026, 6, 1).unwrap());
+        assert_eq!(end, NaiveDate::from_ymd_opt(2026, 7, 1).unwrap());
+
+        // 12月は年をまたいで翌年1月1日になる
+        let (_, end) = parse_year_month_range("2026-12").expect("2026-12 は有効");
+        assert_eq!(end, NaiveDate::from_ymd_opt(2027, 1, 1).unwrap());
+
+        for invalid in ["2026/06", "2026-13", "abcd-06", "2026-06-01"] {
+            assert!(
+                matches!(
+                    parse_year_month_range(invalid),
+                    Err(ApiError::ValidationError(_))
+                ),
+                "value={invalid} は ValidationError になるべき"
+            );
+        }
+    }
+
+    #[test]
+    fn test_parse_date_param_accepts_iso_format_and_rejects_others() {
+        assert!(parse_date_param("date", "2026-01-15").is_ok());
+        for invalid in ["2026/01/15", "15-01-2026", "not-a-date", ""] {
+            assert!(
+                matches!(
+                    parse_date_param("date", invalid),
+                    Err(ApiError::ValidationError(_))
+                ),
+                "value={invalid} は ValidationError になるべき"
+            );
+        }
+    }
+
+    #[test]
+    fn test_escape_like_pattern_escapes_wildcard_characters() {
+        assert_eq!(escape_like_pattern("abc"), "%abc%");
+        assert_eq!(escape_like_pattern("50%"), "%50\\%%");
+        assert_eq!(escape_like_pattern("A_B"), "%A\\_B%");
+        assert_eq!(escape_like_pattern("a\\b"), "%a\\\\b%");
+        assert_eq!(escape_like_pattern("100%_off\\"), "%100\\%\\_off\\\\%");
+    }
 
     #[test]
     fn test_bulk_timer_finish() {
